# openapi.yaml
...
openapi: 3.0.3
info:
  title: "GitHub Insight Engine API"
  description: |
    REST API for GitHub data collection, analysis, and insights generation.
    
    **Authentication**: JWT tokens obtained via `/auth/login`
    **Data Flow**: Collect → Store → Analyze → Retrieve
    **Rate Limits**: Standard GitHub API rate limits apply during collection
    **Automatic Scheduling**: Built-in scheduling service with automatic token management
  version: 0.1.90
  contact:
    name: Github
    url: https://github.com/octocat/gh-insight-engine
  license:
    name: MIT
    url: https://github.com/octocat/gh-insight-engine/blob/main/LICENSE
servers:
  - url: https://github.lucas-blanchard.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
  - {}
paths:
...
  /refresh/{username}:
    post:
      summary: Refresh all user data
      description: |
        Start complete data refresh for a user by chaining all collection steps:
        1. User profile data collection (POST /users/{username})
        2. Repository data collection (POST /repositories/{username})
        3. AI analysis execution (POST /ai/{username})
        
        Returns immediately with status 202 Accepted.
        Use GET /refresh/{username}/status to monitor progress.
        
        This operation can take several minutes. In case of partial failure,
        successfully collected data is preserved.
        
        **Scheduling Integration**: This endpoint is used by the automatic
        scheduling service with automatic token management. The scheduling
        service calls `/auth/login` to obtain fresh JWT tokens before
        executing this refresh operation.
      operationId: refreshUserData
      tags:
        - Refresh
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernameParam'
      responses:
        '202':
          description: Refresh started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshStartedResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "Unauthorized"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "Please log in to access this resource"
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"
        '403':
          description: Can only refresh own data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Complete failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: GitHub API rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "RATE_LIMIT_ERROR"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "GitHub API rate limit exceeded. Please try again later or contact GitHub Support if the issue persists."
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"
  /refresh/{username}/status:
    get:
      summary: Get refresh status
      description: Get the status of data refresh for a user
      operationId: getRefreshStatus
      tags:
        - Refresh
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernameParam'
      responses:
        '200':
          description: Status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshStatusResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "Unauthorized"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "Please log in to access this resource"
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"
        '403':
          description: Can only check own refresh status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No refresh status found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /auth/login (24h validity).
        Used by both manual API calls and the automatic scheduling service.
  parameters:
    UsernameParam:
      name: username
      in: path
      required: true
      description: GitHub username
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]([a-zA-Z0-9\-]){0,38}$'
        example: "octocat"
  schemas:
...
    RefreshStartedResponse:
      type: object
      properties:
        message:
          type: string
          example: "Data refresh started"
        success:
          type: boolean
          example: true
        username:
          type: string
          example: "octocat"
        refreshStatus:
          type: string
          enum: ["in_progress"]
        metadata:
          type: object
          properties:
            startedAt:
              type: string
              format: date-time
            estimatedCompletion:
              type: string
              format: date-time
            currentStep:
              type: integer
              minimum: 1
            totalSteps:
              type: integer
              minimum: 1
        timestamp:
          type: string
          format: date-time

    # Error Response
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          enum:
            - "VALIDATION_ERROR"
            - "AUTHENTICATION_ERROR"
            - "AUTHORIZATION_ERROR"
            - "NOT_FOUND_ERROR"
            - "EXTERNAL_SERVICE_ERROR"
            - "RATE_LIMIT_ERROR"
            - "INTERNAL_SERVER_ERROR"
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "The provided data does not match the expected format"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "githubToken"
              message:
                type: string
                example: "Username is required"
              code:
                type: string
                example: "too_small"
        timestamp:
          type: string
          format: date-time
          example: "2025-07-27T10:18:08.760Z"
        path:
          type: string
          example: "/api/auth/login"
        method:
          type: string
          example: "POST"
        stack:
          type: string
          description: Error stack trace (development only)
...
    RefreshStatusResponse:
      type: object
      properties:
        message:
          type: string
          example: "Refresh in progress"
        status:
          type: string
          enum: ["in_progress", "completed", "failed"]
        progressPercentage:
          type: integer
          minimum: 0
          maximum: 100
          description: Progress percentage (0-100)
        startedAt:
          type: string
          format: date-time
          description: When the refresh started
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: When the refresh completed (if finished)
        estimatedCompletion:
          type: string
          format: date-time
          nullable: true
          description: Estimated completion time
        currentStep:
          type: integer
          minimum: 1
          description: Current step in the refresh process
        totalSteps:
          type: integer
          minimum: 1
          description: Total number of steps in the refresh process
        error:
          type: string
          nullable: true
          description: Error message if refresh failed
        timestamp:
          type: string
          format: date-time
...